bot/requirements.txt
health_ping==1.0.1
python-dotenv==1.1.0
python-telegram-bot==22.1.0
qrcode[pil]==8.2.0
